Move HttpRequestExtensions into more sensible namespace
