+++ Solutions/Menes.PetStore.Hosting/Menes/PetStore/Hosting/DemoOpenApiHost.cs
@@ -5,11 +5,11 @@
 namespace Menes.PetStore.Hosting
 {
     using System.Threading.Tasks;
+    using Menes.Hosting.AspNetCore;
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.Azure.WebJobs;
     using Microsoft.Azure.WebJobs.Extensions.Http;
-    using Microsoft.Extensions.DependencyInjection;
 
     /// <summary>
     /// An example openAPI host.
